SearchScreen 화면 디자인 변경. 색생 등 기본 구성도에서 피그마 디자인으로 변경. 공통된 속성을 갖는 디자인을 Default* 클래스를 통해 모듈화
