+++ app/src/main/java/com/hansung/sherpa/StaticValue.kt
@@ -1,6 +1,5 @@
 package com.hansung.sherpa
 
-import androidx.navigation.NavController
 import com.hansung.sherpa.deviation.RouteControl
 import com.hansung.sherpa.navigation.Navigation
 import com.naver.maps.map.NaverMap
