+++ app/src/main/res/drawable/close.xml
@@ -5,7 +5,7 @@
     android:viewportHeight="96">
     <group>
         <path
-            android:fillColor="#FF000000"
+            android:fillColor="#FFCCCCCC"
             android:pathData="M76 25.64L70.36 20 48 42.36 25.64 20 20 25.64 42.36 48 20 70.36 25.64 76 48 53.64 70.36 76 76 70.36 53.64 48 76 25.64Z"/>
     </group>
 </vector>
