+++ app/src/main/res/drawable/swap.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="35dp"
-    android:height="35dp"
+    android:width="30dp"
+    android:height="30dp"
     android:viewportWidth="96"
     android:viewportHeight="96">
     <path
-        android:fillColor="#FF000000"
+        android:fillColor="#FFCCCCCC"
         android:pathData="M55.89 71.68V47.5c0-1.11 0.38-2.04 1.14-2.8 0.75-0.74 1.7-1.12 2.81-1.12 1.13 0 2.07 0.38 2.83 1.13 0.75 0.75 1.13 1.68 1.13 2.79v24.18l7.42-7.34c0.73-0.72 1.63-1.1 2.72-1.13 1.09-0.03 2.03 0.35 2.82 1.13 0.8 0.72 1.2 1.62 1.24 2.7 0.03 1.07-0.35 2-1.14 2.78L62.6 83.92c-0.4 0.4-0.82 0.67-1.28 0.83-0.46 0.17-0.96 0.25-1.49 0.25-0.52 0-1.02-0.08-1.48-0.25-0.46-0.16-0.89-0.44-1.29-0.83l-14.24-14.1c-0.73-0.71-1.09-1.61-1.09-2.69 0-1.08 0.4-2 1.19-2.79 0.79-0.72 1.71-1.1 2.77-1.13 1.05-0.03 1.98 0.35 2.77 1.13l7.42 7.34ZM32.15 23.32l-7.42 7.34c-0.86 0.85-1.82 1.24-2.87 1.17-1.06-0.06-1.95-0.45-2.67-1.17-0.8-0.78-1.19-1.71-1.19-2.8 0-1.07 0.36-1.97 1.09-2.68l14.24-14.1c0.4-0.4 0.83-0.67 1.29-0.84 0.46-0.16 0.96-0.24 1.48-0.24 0.53 0 1.02 0.08 1.49 0.24 0.46 0.17 0.89 0.45 1.28 0.84l14.25 14.1c0.72 0.71 1.09 1.63 1.09 2.74 0 1.1-0.37 2.02-1.1 2.74-0.78 0.78-1.72 1.17-2.81 1.17s-2.03-0.39-2.82-1.17l-7.42-7.34V47.5c0 1.11-0.38 2.04-1.14 2.79-0.76 0.75-1.7 1.13-2.82 1.13-1.12 0-2.06-0.38-2.81-1.13-0.76-0.75-1.14-1.68-1.14-2.79V23.32Z"/>
 </vector>
